policy todo, e todotask

+++ app/Http/Controllers/TodoController.php
@@ -21,6 +21,8 @@ class TodoController extends Controller
 
     public function show(Todo $todo)
     {
+        $this->authorize('view', $todo);
+
         $todo->load('tasks');
         return new TodoResource($todo);
     }
@@ -35,6 +37,8 @@ class TodoController extends Controller
 
     public function update(Todo $todo, TodoUpdateRequest $request)
     {
+        $this->authorize('udpate', $todo);
+
         $input = $request->validated();
 
         $todo->fill($input);
@@ -45,11 +49,15 @@ class TodoController extends Controller
 
     public function destroy(Todo $todo)
     {
+        $this->authorize('destroy', $todo);
+
         $todo->delete();
     }
 
     public function addTask(Todo $todo, TodoStoreRequest $request)
     {
+        $this->authorize('addTask', $todo);
+
         $input = $request->validated();
 
         $todoTask = $todo->tasks()->create($input);
