+++ app/Http/Controllers/TodoTaskController.php
@@ -16,6 +16,7 @@ class TodoTaskController extends Controller
 
     public function update(TodoTask $todoTask, TodoTaskUpdateRequest $request)
     {
+        $this->authorize('update', $todoTask);
         $input = $request->validated();
 
         $todoTask->fill($input);
@@ -26,6 +27,7 @@ class TodoTaskController extends Controller
 
     public function destroy(TodoTask $todoTask)
     {
+        $this->authorize('destroy', $todoTask);
         $todoTask->delete();
     }
 }
